requirements.txt
requests
python-telegram-bot==13.20